// part9/Patientor/backend/types.ts
export enum Gender {
  Male = 'male',
  Female = 'female',
  Other = 'other',
}

export interface Diagnose {
  code: string;
  name: string;
  latin?: string;
}

export interface Patient {
  id: string;
  name: string;
  dateOfBirth: string;
  ssn?: string;
  gender: Gender;
  occupation: string;
  entries: Entry[];
}

export type NewPatient = Omit<Patient, 'id' | 'entries'>;

export type PublicPatient = Omit<Patient, 'ssn' | 'entries'>;

interface BaseEntry {
  id: string;
  description: string;
  date: string;
  specialist: string;
  diagnosisCodes?: Array<Diagnose['code']>;
}

export enum HealthCheckRating {
  "Healthy" = 0,
  "LowRisk" = 1,
  "HighRisk" = 2,
  "CriticalRisk" = 3
}

interface HealthCheckEntry extends BaseEntry {
  type: "HealthCheck";
  healthCheckRating: HealthCheckRating;
}

export interface HospitalEntry extends BaseEntry {
  type:"Hospital";
  discharge: {date: string, criteria: string}
}

export interface OccupationalHealthcareEntry extends BaseEntry {
  type: "OccupationalHealthcare";
  employerName: string;
  sickLeave?: {startDate: string, endDate: string}
}

export type typeEntries = "HealthCheck" | "Hospital" | "OccupationalHealthcare";

export type Entry =
  | HospitalEntry
  | OccupationalHealthcareEntry
  | HealthCheckEntry;

type InputHealthCheckEntry = Omit<HealthCheckEntry, 'id'>;
type InputHospitalEntry = Omit<HospitalEntry, 'id'>;
type InputOccupationalHealthcareEntry = Omit<OccupationalHealthcareEntry, 'id'>;

export type InputEntry = 
| InputHospitalEntry
| InputOccupationalHealthcareEntry
| InputHealthCheckEntry;